mcp3008: Tidy up and fix auto load bug
