+++ esphome/components/mcp3008/mcp3008.cpp
@@ -1,4 +1,6 @@
 #include "mcp3008.h"
+
+#include "esphome/core/helpers.h"
 #include "esphome/core/log.h"
 
 namespace esphome {
@@ -32,28 +34,10 @@ float MCP3008::read_data(uint8_t pin) {
 
   this->disable();
 
-  int data = data_msb << 8 | data_lsb;
+  uint16_t data = encode_uint16(data_msb, data_lsb);
 
   return data / 1023.0f;
 }
-
-MCP3008Sensor::MCP3008Sensor(MCP3008 *parent, uint8_t pin, float reference_voltage)
-    : PollingComponent(1000), parent_(parent), pin_(pin), reference_voltage_(reference_voltage) {}
-
-float MCP3008Sensor::get_setup_priority() const { return setup_priority::DATA; }
-
-void MCP3008Sensor::setup() { LOG_SENSOR("", "Setting up MCP3008 Sensor '%s'...", this); }
-void MCP3008Sensor::dump_config() {
-  ESP_LOGCONFIG(TAG, "MCP3008Sensor:");
-  ESP_LOGCONFIG(TAG, "  Pin: %u", this->pin_);
-  ESP_LOGCONFIG(TAG, "  Reference Voltage: %.2fV", this->reference_voltage_);
-}
-float MCP3008Sensor::sample() {
-  float value_v = this->parent_->read_data(pin_);
-  value_v = (value_v * this->reference_voltage_);
-  return value_v;
-}
-void MCP3008Sensor::update() { this->publish_state(this->sample()); }
 
 }  // namespace mcp3008
 }  // namespace esphome
